Port many tests using legacy matchers to munit
