+++ project/Http4sPlugin.scala
@@ -356,6 +356,7 @@ object Http4sPlugin extends AutoPlugin {
     val reactiveStreams = "1.0.3"
     val quasiquotes = "2.1.0"
     val scalacheck = "1.15.1"
+    val scalacheckEffect = "0.6.0"
     val scalafix = _root_.scalafix.sbt.BuildInfo.scalafixVersion
     val scalatags = "0.9.2"
     val scalaXml = "1.3.0"
@@ -431,6 +432,8 @@ object Http4sPlugin extends AutoPlugin {
   lazy val reactiveStreams                  = "org.reactivestreams"    %  "reactive-streams"          % V.reactiveStreams
   lazy val quasiquotes                      = "org.scalamacros"        %% "quasiquotes"               % V.quasiquotes
   lazy val scalacheck                       = "org.scalacheck"         %% "scalacheck"                % V.scalacheck
+  lazy val scalacheckEffect                 = "org.typelevel"          %% "scalacheck-effect"         % V.scalacheckEffect
+  lazy val scalacheckEffectMunit            = "org.typelevel"          %% "scalacheck-effect-munit"   % V.scalacheckEffect
   def scalaReflect(sv: String)              = "org.scala-lang"         %  "scala-reflect"             % sv
   lazy val scalatagsApi                     = "com.lihaoyi"            %% "scalatags"                 % V.scalatags
   lazy val scalaXml                         = "org.scala-lang.modules" %% "scala-xml"                 % V.scalaXml
