+++ server/src/test/scala/org/http4s/server/middleware/EntityLimiterSuite.scala
@@ -1,0 +1,85 @@
+/*
+ * Copyright 2013-2020 http4s.org
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package org.http4s
+package server
+package middleware
+
+import cats.effect._
+import cats.implicits._
+import fs2._
+import fs2.Stream._
+import java.nio.charset.StandardCharsets
+import org.http4s.Method._
+import org.http4s.Status._
+import org.http4s.Uri.uri
+import org.http4s.syntax.all._
+import org.http4s.server.middleware.EntityLimiter.EntityTooLarge
+
+class EntityLimiterSuite extends Http4sSuite {
+  val routes = HttpRoutes.of[IO] {
+    case r if r.uri.path == "/echo" => r.decode[String](Response[IO](Ok).withEntity(_).pure[IO])
+  }
+
+  val b = chunk(Chunk.bytes("hello".getBytes(StandardCharsets.UTF_8)))
+
+  test("Allow reasonable entities") {
+    EntityLimiter(routes, 100)
+      .apply(Request[IO](POST, uri("/echo"), body = b))
+      .map(_ => -1)
+      .value
+      .assertEquals(Some(-1))
+  }
+
+  test("Limit the maximum size of an EntityBody") {
+    EntityLimiter(routes, 3)
+      .apply(Request[IO](POST, uri("/echo"), body = b))
+      .map(_ => -1L)
+      .value
+      .handleError { case EntityTooLarge(i) => Some(i) }
+      .assertEquals(Some(3L))
+  }
+
+  test("Chain correctly with other HttpRoutes") {
+    val routes2 = HttpRoutes.of[IO] {
+      case r if r.uri.path == "/echo2" =>
+        r.decode[String](Response[IO](Ok).withEntity(_).pure[IO])
+    }
+
+    val st = EntityLimiter(routes, 3) <+> routes2
+
+    st.apply(Request[IO](POST, uri("/echo2"), body = b))
+      .map(_ => -1)
+      .value
+      .assertEquals(Some(-1)) *>
+      st.apply(Request[IO](POST, uri("/echo"), body = b))
+        .map(_ => -1L)
+        .value
+        .handleError { case EntityTooLarge(i) => Some(i) }
+        .assertEquals(Some(3L))
+  }
+
+  test("Be created via the httpRoutes constructor") {
+    EntityLimiter
+      .httpRoutes(routes, 3)
+      .apply(Request[IO](POST, uri("/echo"), body = b))
+      .map(_ => -1L)
+      .value
+      .handleError { case EntityTooLarge(i) => Some(i) }
+      .assertEquals(Some(3L))
+  }
+
+  test("Be created via the httpRoutes constructor") {
+    val app: HttpApp[IO] = routes.orNotFound
+
+    EntityLimiter
+      .httpApp(app, 3L)
+      .apply(Request[IO](POST, uri("/echo"), body = b))
+      .map(_ => -1L)
+      .handleError { case EntityTooLarge(i) => i }
+      .assertEquals(3L)
+  }
+}
