+++ server/src/test/scala/org/http4s/server/middleware/HttpsRedirectSuite.scala
@@ -16,7 +16,7 @@ import org.http4s.headers._
 import org.http4s.syntax.all._
 import org.http4s.Uri.{Authority, RegName, Scheme}
 
-class HttpsRedirectSpec extends Http4sSuite {
+class HttpsRedirectSuite extends Http4sSuite {
   val innerRoutes = HttpRoutes.of[IO] { case GET -> Root =>
     Ok("pong")
   }
