+++ server/src/test/scala/org/http4s/server/middleware/MaxActiveRequestsSuite.scala
@@ -1,0 +1,88 @@
+/*
+ * Copyright 2013-2020 http4s.org
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package org.http4s.server.middleware
+
+import cats.implicits._
+import cats.effect._
+import cats.data._
+import cats.effect.concurrent._
+import org.http4s._
+import org.http4s.syntax.all._
+
+class MaxActiveRequestsSuite extends Http4sSuite {
+  val req = Request[IO]()
+
+  def routes(startedGate: Deferred[IO, Unit], deferred: Deferred[IO, Unit]) =
+    Kleisli { (req: Request[IO]) =>
+      req match {
+        case other if other.method == Method.PUT => OptionT.none[IO, Response[IO]]
+        case _ =>
+          OptionT.liftF(
+            startedGate.complete(()) >> deferred.get >> Response[IO](Status.Ok).pure[IO])
+      }
+    }
+
+  test("httpApp allow a request when allowed") {
+    (for {
+      deferredStarted <- Deferred[IO, Unit]
+      deferredWait <- Deferred[IO, Unit]
+      _ <- deferredWait.complete(())
+      middle <- MaxActiveRequests.httpApp[IO](1)
+      httpApp = middle(routes(deferredStarted, deferredWait).orNotFound)
+      out <- httpApp.run(req)
+    } yield out.status).assertEquals(Status.Ok)
+  }
+
+  test("httpApp not allow a request if max active") {
+    (for {
+      deferredStarted <- Deferred[IO, Unit]
+      deferredWait <- Deferred[IO, Unit]
+      middle <- MaxActiveRequests.httpApp[IO](1)
+      httpApp = middle(routes(deferredStarted, deferredWait).orNotFound)
+      f <- httpApp.run(req).start
+      _ <- deferredStarted.get
+      out <- httpApp.run(req)
+      _ <- f.cancel
+    } yield out.status).assertEquals(Status.ServiceUnavailable)
+
+  }
+  test("httpRoutes allow a request when allowed") {
+    (for {
+      deferredStarted <- Deferred[IO, Unit]
+      deferredWait <- Deferred[IO, Unit]
+      _ <- deferredWait.complete(())
+      middle <- MaxActiveRequests.httpRoutes[IO](1)
+      httpApp = middle(routes(deferredStarted, deferredWait)).orNotFound
+      out <- httpApp.run(req)
+    } yield out.status).assertEquals(Status.Ok)
+  }
+
+  test("httpRoutes not allow a request if max active") {
+    (for {
+      deferredStarted <- Deferred[IO, Unit]
+      deferredWait <- Deferred[IO, Unit]
+      middle <- MaxActiveRequests.httpRoutes[IO](1)
+      httpApp = middle(routes(deferredStarted, deferredWait)).orNotFound
+      f <- httpApp.run(req).start
+      _ <- deferredStarted.get
+      out <- httpApp.run(req)
+      _ <- f.cancel
+    } yield out.status).assertEquals(Status.ServiceUnavailable)
+  }
+
+  test("httpRoutes release resource on None") {
+    (for {
+      deferredStarted <- Deferred[IO, Unit]
+      deferredWait <- Deferred[IO, Unit]
+      middle <- MaxActiveRequests.httpRoutes[IO](1)
+      httpApp = middle(routes(deferredStarted, deferredWait)).orNotFound
+      out1 <- httpApp.run(Request(Method.PUT))
+      _ <- deferredWait.complete(())
+      out2 <- httpApp.run(req)
+    } yield (out1.status, out2.status)).assertEquals((Status.NotFound, Status.Ok))
+  }
+}
