+++ testing/src/test/scala/org/http4s/Http4sSuite.scala
@@ -6,10 +6,18 @@
 
 package org.http4s
 
+import cats.effect.Blocker
+import org.http4s.internal.threads.newBlockingPool
 import munit._
 
 /** Common stack for http4s' munit based tests
   */
-trait Http4sSuite extends CatsEffectSuite with DisciplineSuite {}
+trait Http4sSuite extends CatsEffectSuite with DisciplineSuite with munit.ScalaCheckEffectSuite {
 
-object Http4sSuite {}
+  val testBlocker: Blocker = Http4sSpec.TestBlocker
+}
+
+object Http4sSuite {
+  val TestBlocker: Blocker =
+    Blocker.liftExecutorService(newBlockingPool("http4s-spec-blocking"))
+}
